@@ -1,53 +1,37 @@
 class FundsController < ApplicationController
+  before_action :set_fund, only: [:show, :edit, :update, :destroy]
+
   # GET /funds
   # GET /funds.json
   def index
     @funds = Fund.all
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.json { render json: @funds }
-    end
   end
 
   # GET /funds/1
   # GET /funds/1.json
   def show
-    @fund = Fund.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.json { render json: @fund }
-    end
   end
 
   # GET /funds/new
-  # GET /funds/new.json
   def new
     @fund = Fund.new
-
-    respond_to do |format|
-      format.html # new.html.erb
-      format.json { render json: @fund }
-    end
   end
 
   # GET /funds/1/edit
   def edit
-    @fund = Fund.find(params[:id])
   end
 
   # POST /funds
   # POST /funds.json
   def create
     @fund = Fund.new(fund_params)
-    
+    @fund.currency = Currency.find_by!(sym: :RMB)
     respond_to do |format|
       if @fund.save
         format.html { redirect_to @fund, notice: 'Fund was successfully created.' }
-        format.json { render json: @fund, status: :created, location: @fund }
+        format.json { render :show, status: :created, location: @fund }
       else
-        format.html { render action: "new" }
+        format.html { render :new }
         format.json { render json: @fund.errors, status: :unprocessable_entity }
       end
     end
@@ -56,14 +40,12 @@ class FundsController < ApplicationController
   # PATCH/PUT /funds/1
   # PATCH/PUT /funds/1.json
   def update
-    @fund = Fund.find(params[:id])
-
     respond_to do |format|
-      if @fund.update_attributes(fund_params)
+      if @fund.update(fund_params)
         format.html { redirect_to @fund, notice: 'Fund was successfully updated.' }
-        format.json { head :no_content }
+        format.json { render :show, status: :ok, location: @fund }
       else
-        format.html { render action: "edit" }
+        format.html { render :edit }
         format.json { render json: @fund.errors, status: :unprocessable_entity }
       end
     end
@@ -72,21 +54,21 @@ class FundsController < ApplicationController
   # DELETE /funds/1
   # DELETE /funds/1.json
   def destroy
-    @fund = Fund.find(params[:id])
     @fund.destroy
-
     respond_to do |format|
-      format.html { redirect_to funds_url }
+      format.html { redirect_to funds_url, notice: 'Fund was successfully destroyed.' }
       format.json { head :no_content }
     end
   end
 
   private
+    # Use callbacks to share common setup or constraints between actions.
+    def set_fund
+      @fund = Fund.find(params[:id])
+    end
 
-    # Use this method to whitelist the permissible parameters. Example:
-    # params.require(:person).permit(:name, :age)
-    # Also, you can specialize this method with per-user checking of permissible attributes.
+    # Never trust parameters from the scary internet, only allow the white list through.
     def fund_params
-      params.require(:fund).permit(:amount, :comment, :fund_type_id, :time)
+      params.require(:fund).permit(:amount, :time, :fund_type_id, :comment, :currency_id, :origin_amount)
     end
 end
