@@ -1,40 +1,24 @@
 class ProjectTypesController < ApplicationController
+  before_action :set_project_type, only: [:show, :edit, :update, :destroy]
+
   # GET /project_types
   # GET /project_types.json
   def index
     @project_types = ProjectType.all
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.json { render json: @project_types }
-    end
   end
 
   # GET /project_types/1
   # GET /project_types/1.json
   def show
-    @project_type = ProjectType.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.json { render json: @project_type }
-    end
   end
 
   # GET /project_types/new
-  # GET /project_types/new.json
   def new
     @project_type = ProjectType.new
-
-    respond_to do |format|
-      format.html # new.html.erb
-      format.json { render json: @project_type }
-    end
   end
 
   # GET /project_types/1/edit
   def edit
-    @project_type = ProjectType.find(params[:id])
   end
 
   # POST /project_types
@@ -45,9 +29,9 @@ class ProjectTypesController < ApplicationController
     respond_to do |format|
       if @project_type.save
         format.html { redirect_to @project_type, notice: 'Project type was successfully created.' }
-        format.json { render json: @project_type, status: :created, location: @project_type }
+        format.json { render :show, status: :created, location: @project_type }
       else
-        format.html { render action: "new" }
+        format.html { render :new }
         format.json { render json: @project_type.errors, status: :unprocessable_entity }
       end
     end
@@ -56,14 +40,12 @@ class ProjectTypesController < ApplicationController
   # PATCH/PUT /project_types/1
   # PATCH/PUT /project_types/1.json
   def update
-    @project_type = ProjectType.find(params[:id])
-
     respond_to do |format|
-      if @project_type.update_attributes(project_type_params)
+      if @project_type.update(project_type_params)
         format.html { redirect_to @project_type, notice: 'Project type was successfully updated.' }
-        format.json { head :no_content }
+        format.json { render :show, status: :ok, location: @project_type }
       else
-        format.html { render action: "edit" }
+        format.html { render :edit }
         format.json { render json: @project_type.errors, status: :unprocessable_entity }
       end
     end
@@ -72,20 +54,20 @@ class ProjectTypesController < ApplicationController
   # DELETE /project_types/1
   # DELETE /project_types/1.json
   def destroy
-    @project_type = ProjectType.find(params[:id])
     @project_type.destroy
-
     respond_to do |format|
-      format.html { redirect_to project_types_url }
+      format.html { redirect_to project_types_url, notice: 'Project type was successfully destroyed.' }
       format.json { head :no_content }
     end
   end
 
   private
+    # Use callbacks to share common setup or constraints between actions.
+    def set_project_type
+      @project_type = ProjectType.find(params[:id])
+    end
 
-    # Use this method to whitelist the permissible parameters. Example:
-    # params.require(:person).permit(:name, :age)
-    # Also, you can specialize this method with per-user checking of permissible attributes.
+    # Never trust parameters from the scary internet, only allow the white list through.
     def project_type_params
       params.require(:project_type).permit(:name)
     end
