@@ -1,40 +1,24 @@
 class UnivUnitManagersController < ApplicationController
+  before_action :set_univ_unit_manager, only: [:show, :edit, :update, :destroy]
+
   # GET /univ_unit_managers
   # GET /univ_unit_managers.json
   def index
     @univ_unit_managers = UnivUnitManager.all
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.json { render json: @univ_unit_managers }
-    end
   end
 
   # GET /univ_unit_managers/1
   # GET /univ_unit_managers/1.json
   def show
-    @univ_unit_manager = UnivUnitManager.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.json { render json: @univ_unit_manager }
-    end
   end
 
   # GET /univ_unit_managers/new
-  # GET /univ_unit_managers/new.json
   def new
     @univ_unit_manager = UnivUnitManager.new
-
-    respond_to do |format|
-      format.html # new.html.erb
-      format.json { render json: @univ_unit_manager }
-    end
   end
 
   # GET /univ_unit_managers/1/edit
   def edit
-    @univ_unit_manager = UnivUnitManager.find(params[:id])
   end
 
   # POST /univ_unit_managers
@@ -45,9 +29,9 @@ class UnivUnitManagersController < ApplicationController
     respond_to do |format|
       if @univ_unit_manager.save
         format.html { redirect_to @univ_unit_manager, notice: 'Univ unit manager was successfully created.' }
-        format.json { render json: @univ_unit_manager, status: :created, location: @univ_unit_manager }
+        format.json { render :show, status: :created, location: @univ_unit_manager }
       else
-        format.html { render action: "new" }
+        format.html { render :new }
         format.json { render json: @univ_unit_manager.errors, status: :unprocessable_entity }
       end
     end
@@ -56,14 +40,12 @@ class UnivUnitManagersController < ApplicationController
   # PATCH/PUT /univ_unit_managers/1
   # PATCH/PUT /univ_unit_managers/1.json
   def update
-    @univ_unit_manager = UnivUnitManager.find(params[:id])
-
     respond_to do |format|
-      if @univ_unit_manager.update_attributes(univ_unit_manager_params)
+      if @univ_unit_manager.update(univ_unit_manager_params)
         format.html { redirect_to @univ_unit_manager, notice: 'Univ unit manager was successfully updated.' }
-        format.json { head :no_content }
+        format.json { render :show, status: :ok, location: @univ_unit_manager }
       else
-        format.html { render action: "edit" }
+        format.html { render :edit }
         format.json { render json: @univ_unit_manager.errors, status: :unprocessable_entity }
       end
     end
@@ -72,20 +54,20 @@ class UnivUnitManagersController < ApplicationController
   # DELETE /univ_unit_managers/1
   # DELETE /univ_unit_managers/1.json
   def destroy
-    @univ_unit_manager = UnivUnitManager.find(params[:id])
     @univ_unit_manager.destroy
-
     respond_to do |format|
-      format.html { redirect_to univ_unit_managers_url }
+      format.html { redirect_to univ_unit_managers_url, notice: 'Univ unit manager was successfully destroyed.' }
       format.json { head :no_content }
     end
   end
 
   private
+    # Use callbacks to share common setup or constraints between actions.
+    def set_univ_unit_manager
+      @univ_unit_manager = UnivUnitManager.find(params[:id])
+    end
 
-    # Use this method to whitelist the permissible parameters. Example:
-    # params.require(:person).permit(:name, :age)
-    # Also, you can specialize this method with per-user checking of permissible attributes.
+    # Never trust parameters from the scary internet, only allow the white list through.
     def univ_unit_manager_params
       params.require(:univ_unit_manager).permit(:name, :univ_unit_id)
     end
