@@ -1,40 +1,26 @@
 class UsageRecordsController < ApplicationController
+  before_action :set_usage_record, only: [:show, :edit, :update, :destroy]
+  before_action :set_project, only: [:new, :create, :edit, :update, :destroy]
+
   # GET /usage_records
   # GET /usage_records.json
   def index
     @usage_records = UsageRecord.all
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.json { render json: @usage_records }
-    end
   end
 
   # GET /usage_records/1
   # GET /usage_records/1.json
   def show
-    @usage_record = UsageRecord.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.json { render json: @usage_record }
-    end
   end
 
   # GET /usage_records/new
-  # GET /usage_records/new.json
   def new
     @usage_record = UsageRecord.new
-    @usage_record.project_id = params[:project]
-    respond_to do |format|
-      format.html # new.html.erb
-      format.json { render json: @usage_record }
-    end
+    @usage_record.project = Project.find(params[:project_id])
   end
 
   # GET /usage_records/1/edit
   def edit
-    @usage_record = UsageRecord.find(params[:id])
   end
 
   # POST /usage_records
@@ -44,10 +30,10 @@ class UsageRecordsController < ApplicationController
 
     respond_to do |format|
       if @usage_record.save
-        format.html { redirect_to project_path(@usage_record.project_id), notice: 'Usage record was successfully created.' }
-        format.json { render json: @usage_record, status: :created, location: @usage_record }
+        format.html { redirect_to @project, notice: 'Usage record was successfully created.' }
+        format.json { render :show, status: :created, location: @usage_record }
       else
-        format.html { render action: "new" }
+        format.html { render :new }
         format.json { render json: @usage_record.errors, status: :unprocessable_entity }
       end
     end
@@ -56,14 +42,12 @@ class UsageRecordsController < ApplicationController
   # PATCH/PUT /usage_records/1
   # PATCH/PUT /usage_records/1.json
   def update
-    @usage_record = UsageRecord.find(params[:id])
-
     respond_to do |format|
-      if @usage_record.update_attributes(usage_record_params)
-        format.html { redirect_to @usage_record, notice: 'Usage record was successfully updated.' }
-        format.json { head :no_content }
+      if @usage_record.update(usage_record_params)
+        format.html { redirect_to @project, notice: 'Usage record was successfully updated.' }
+        format.json { render :show, status: :ok, location: @usage_record }
       else
-        format.html { render action: "edit" }
+        format.html { render :edit }
         format.json { render json: @usage_record.errors, status: :unprocessable_entity }
       end
     end
@@ -72,28 +56,37 @@ class UsageRecordsController < ApplicationController
   # DELETE /usage_records/1
   # DELETE /usage_records/1.json
   def destroy
-    @usage_record = UsageRecord.find(params[:id])
     @usage_record.destroy
-
     respond_to do |format|
-      format.html { redirect_to usage_records_url }
+      format.html { redirect_to @project, notice: 'Usage record was successfully destroyed.' }
       format.json { head :no_content }
     end
   end
 
   private
-
-    # Use this method to whitelist the permissible parameters. Example:
-    # params.require(:person).permit(:name, :age)
-    # Also, you can specialize this method with per-user checking of permissible attributes.
+    # Use callbacks to share common setup or constraints between actions.
+    def set_usage_record
+      @usage_record = UsageRecord.find(params[:id])
+    end
+    def set_project
+      @project =  (
+        if @usage_record && @usage_record.project
+          @usage_record.project
+        else
+          Project.find(params[:project_id])
+        end
+      )
+    end
+    # Never trust parameters from the scary internet, only allow the white list through.
     def usage_record_params
       params.require(:usage_record)
-        .permit(  :customer_id, :project_id, :exec_unit_id, :exec_manager_id, 
-                  :benefit_unit_id, :benefit_manager_id,:usage_type_id,
-                  fund_attributes: [:amount, :time, :fund_type_id]
-                  )
-        .tap{ |p| 
-          p[:employee_id] = @current_user.id
-        }
+      .permit(:customer_id, :project_id, :exec_unit_id, :exec_manager_id, 
+              :benefit_unit_id, :benefit_manager_id,:usage_type_id,
+              fund_attributes: [:amount, :time, :fund_type_id]
+              )
+        .tap{ |p|
+        p[:creator_id] = current_people.id
+        p[:project_id] = params[:project_id]
+      }
     end
 end
