@@ -1,13 +1,16 @@
 class User < ActiveRecord::Base
-  belongs_to :user, polymorphic: true, dependent: :destroy
-  
-  attr_accessible :account, :password, :password_confirmation, :user_id
-  
-  has_secure_password
+  include PolymorphicExtension
+  # Include default devise modules. Others available are:
+  # :confirmable, :lockable, :timeoutable and :omniauthable
+  devise :database_authenticatable, :registerable,
+         :recoverable, :rememberable, :trackable, :validatable
 
-  validates_presence_of :password, :password_confirmation # validate confirmation will not work if password_fonfirmation is nil! 
-                                                          # Don't forget to check your patams permit in controller too.
-  validates_confirmation_of :password
-  validates :account, :presence => true, :uniqueness => true
-  validates :password, :presence => true
+  belongs_to :user, polymorphic: true
+  validates :account, uniqueness: true, presence: true
+
+  protected
+
+    def email_required?
+      false
+    end
 end
