@@ -10,10 +10,12 @@ class CustomersController < ApplicationController
   # GET /customers/1
   # GET /customers/1.json
   def show
+    redirect_to @customer.customer
   end
 
   # GET /customers/new
   def new
+    @customer = Customer.new
   end
 
   # GET /customers/1/edit
@@ -28,9 +30,9 @@ class CustomersController < ApplicationController
     respond_to do |format|
       if @customer.save
         format.html { redirect_to @customer, notice: 'Customer was successfully created.' }
-        format.json { render action: 'show', status: :created, location: @customer }
+        format.json { render :show, status: :created, location: @customer }
       else
-        format.html { render action: 'new' }
+        format.html { render :new }
         format.json { render json: @customer.errors, status: :unprocessable_entity }
       end
     end
@@ -42,9 +44,9 @@ class CustomersController < ApplicationController
     respond_to do |format|
       if @customer.update(customer_params)
         format.html { redirect_to @customer, notice: 'Customer was successfully updated.' }
-        format.json { head :no_content }
+        format.json { render :show, status: :ok, location: @customer }
       else
-        format.html { render action: 'edit' }
+        format.html { render :edit }
         format.json { render json: @customer.errors, status: :unprocessable_entity }
       end
     end
@@ -55,7 +57,7 @@ class CustomersController < ApplicationController
   def destroy
     @customer.destroy
     respond_to do |format|
-      format.html { redirect_to customers_url }
+      format.html { redirect_to customers_url, notice: 'Customer was successfully destroyed.' }
       format.json { head :no_content }
     end
   end
@@ -68,6 +70,6 @@ class CustomersController < ApplicationController
 
     # Never trust parameters from the scary internet, only allow the white list through.
     def customer_params
-      params.require(:customer).permit(:name, :customer_type_id, :employee_id)
+      params.require(:customer).permit(:name, :customer_id, :customer_type)
     end
 end
