@@ -1,3 +1,3 @@
 class Currency < ActiveRecord::Base
-  attr_accessible :name
+  has_many :funds
 end
