@@ -1,11 +1,18 @@
 class Customer < ActiveRecord::Base
+  include PolymorphicExtension
   belongs_to :customer, polymorphic: true, dependent: :destroy
-  has_and_belongs_to_many :customer_groups
-  # belongs_to :customer_type
-  belongs_to :employee
+  # has_and_belongs_to_many :customer_groups
+  belongs_to :creator, class_name: "Employee"
   has_many :donation_records
-  attr_accessible :name, :employee_id
+  has_many :contact_records
+
+  has_many :project_link_men
+  has_many :link_projects, class_name: :Project, through: :project_link_men, source: :project
+
   validates_presence_of :name
+  
+  set_accessable_attributes [:total_donation, :contact_records, :link_projects]
+
   def customer_type
     super
   end
@@ -15,11 +22,6 @@ class Customer < ActiveRecord::Base
       amount+=r.plan_fund.amount
     end
     amount
-  end
-  def self.accessable_attributes
-    #don't use << to append!!
-    #see more: https://ruby-china.org/topics/10084
-    self.column_names + ["total_donation"]
   end
   def to_s
     name
