@@ -1,6 +1,6 @@
 class UsageRecord < ActiveRecord::Base
   belongs_to :fund, dependent: :destroy
-  belongs_to :employee
+  belongs_to :creator, class_name: :Employee, foreign_key: :creator_id
   belongs_to :project
   belongs_to :exec_unit, class_name: :UnivUnit
   belongs_to :exec_manager, class_name: :UnivUnitManager
@@ -10,8 +10,9 @@ class UsageRecord < ActiveRecord::Base
 
   validates_associated :fund
 
+  validates_presence_of_all except: :fund_id
+
   accepts_nested_attributes_for :fund
   auto_build :fund
 
-  # attr_accessible :title, :body
 end
